Move Gamepad structs to their respective correct location - Structs that are reversed from the Game are moved to Structs.hpp - Structs that are custom to the Gamepad patch are moved into the Gamepad component class - Offsets to the Game's memory are moved to the Functions.hpp header - Arrays that hold data are moved to the Gamepad components class as static fields
